refactor: Menu 엔티티를 별도의 도메인 패키지로 이동
