+++ src/main/java/com/jvnlee/catchdining/domain/restaurant/model/Restaurant.java
@@ -1,5 +1,6 @@
 package com.jvnlee.catchdining.domain.restaurant.model;
 
+import com.jvnlee.catchdining.domain.menu.domain.Menu;
 import com.jvnlee.catchdining.domain.restaurant.dto.RestaurantDto;
 import com.jvnlee.catchdining.domain.seat.model.Seat;
 import com.jvnlee.catchdining.entity.*;
